Introduce react sagas and nuke everything
