+++ client/resources/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="theme-color" content="#000000">
 
-    <title>Money Mile | A Lyft API Application</title>
+    <title>purlieu.io</title>
 
     <link rel="stylesheet" href="./css/main.css"></link>
     <link href="https://fonts.googleapis.com/css?family=Yantramanav|Source+Code+Pro" rel="stylesheet">
