+++ client/store/index.js
@@ -1,0 +1,1 @@
+// mount it on the Store
