+++ package.json
@@ -45,6 +45,7 @@
     "eslint-plugin-react": "^7.11.1",
     "prettier": "^1.15.3",
     "react-redux": "^6.0.0",
+    "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
     "redux-saga": "^0.16.2",
     "webpack": "^4.0.0",
